Add main_compute_benchmark.cc to act as main function for benchmarks requiring initialization. Add new ADD_ARROW_COMPUTE_BENCHMARK function to add the new source to the compilation unit and link arrow_compute
