+++ cpp/src/arrow/compute/kernels/scalar_arithmetic_benchmark.cc
@@ -22,7 +22,6 @@
 #include "arrow/array.h"
 #include "arrow/chunked_array.h"
 #include "arrow/compute/api_scalar.h"
-#include "arrow/compute/initialize.h"
 #include "arrow/datum.h"
 #include "arrow/testing/gtest_util.h"
 #include "arrow/testing/random.h"
@@ -162,15 +161,3 @@ DECLARE_ARITHMETIC_CHECKED_BENCHMARKS(ArrayScalarKernel, DivideChecked);
 
 }  // namespace compute
 }  // namespace arrow
-
-int main(int argc, char** argv) {
-  // Initialize compute functions before any benchmarks run
-  ABORT_NOT_OK(arrow::compute::Initialize());
-
-  // Initialize and run benchmarks
-  ::benchmark::Initialize(&argc, argv);
-  if (::benchmark::ReportUnrecognizedArguments(argc, argv)) return 1;
-  ::benchmark::RunSpecifiedBenchmarks();
-  ::benchmark::Shutdown();
-  return 0;
-}
